+++ lib/membrane/core/element/demand_handler.ex
@@ -28,9 +28,8 @@ defmodule Membrane.Core.Element.DemandHandler do
   Called when redemand action was returned.
     * If element is currently supplying demand it means that after finishing supply_demand it will call
       `handle_delayed_demands`.
-    * If element isn't supplying demand at the moment `handle_demand` is invoked right away, and it will
-      invoke handle_demand callback, which will probably return :redemand and :buffers actions and in
-      that way source will synchronously supply demand.
+    * If <span class="x x-first x-last">the </span>element isn't supplying demand at the moment <span class="x x-first x-last">and there's some unsupplied demand on the given </span>
+      <span class="x x-first x-last">output, `</span>handle_demand<span class="x x-first x-last">` is invoked right away, so that the demand can be synchronously supplied.</span>
   """
   @spec handle_redemand(Pad.ref_t(), State.t()) :: {:ok, State.t()}
   def handle_redemand(pad_ref, %State{supplying_demand?: true} = state) do